requirements.txt
langchain>=0.3
langchain-text-splitters>=0.3
langchain-community>=0.3
langchain-chroma>=0.2
langchain-ollama>=0.3
langchain-google-genai>=2.0
chromadb>=0.5
pypdf>=4.0
python-dotenv>=1.0
langsmith>=0.2
